测试代码error优化

+++ src/controllers/oms/jobOrder/JobOrderService.java
@@ -306,6 +306,9 @@ public class JobOrderService extends Controller {
         
         JobOrder jobOrder = new JobOrder();
         
+        if(type == null){
+        	type = "";
+        }
         //获取office_id
    		UserLogin user = LoginUserController.getLoginUser(this);
    		if(user==null){
@@ -333,7 +336,7 @@ public class JobOrderService extends Controller {
    			//update
    			jobOrder = JobOrder.dao.findById(id);
    			//版本(时间戳)校验，不对的话就不让更新保存
-   			Timestamp page_update_stamp = Timestamp.valueOf(dto.get("update_stamp").toString());
+   			Timestamp page_update_stamp = Timestamp.valueOf((String)dto.get("update_stamp"));
    			Timestamp order_update_stamp = jobOrder.getTimestamp("update_stamp");
    			if(!order_update_stamp.equals(page_update_stamp)){
    			    Record rec = new Record();
@@ -427,7 +430,7 @@ public class JobOrderService extends Controller {
    			}
    			
    		}
-   		long customerId = Long.valueOf(dto.get("customer_id").toString());
+   		long customerId = Long.valueOf((String)dto.get("customer_id"));
    		//常用客户保存进入历史记录
    		saveCustomerQueryHistory(customerId);
 		//海运
@@ -561,30 +564,36 @@ public class JobOrderService extends Controller {
    		saveAirTemplate(air_detail);
    		
    		//海 陆 空运带出合同费用信息(供应商)
-   		if(supplier_contract_type.indexOf("ocean")>-1){
-   			if(type.indexOf("柜货")>-1){
-   	   			saveJobOceanGHSpContractConditions(jobOrder);
-   	   		}else if(type.indexOf("散货")>-1){
-   	   			saveJobOceanSHSpContractConditions(jobOrder);
+   		if(supplier_contract_type != null){
+   			if(supplier_contract_type.indexOf("ocean")>-1){
+   	   			if(type.indexOf("柜货")>-1){
+   	   	   			saveJobOceanGHSpContractConditions(jobOrder);
+   	   	   		}else if(type.indexOf("散货")>-1){
+   	   	   			saveJobOceanSHSpContractConditions(jobOrder);
+   	   	   		}
+   	   		}
+   	   		if(supplier_contract_type.indexOf("air")>-1){
+   	   			saveJobAirSpContractConditions(jobOrder);
+   	   		}
+   	   		if(supplier_contract_type.indexOf("land")>-1){
+   	   			saveJobLandSpContractConditions(jobOrder);
    	   		}
    		}
-   		if(supplier_contract_type.indexOf("air")>-1){
-   			saveJobAirSpContractConditions(jobOrder);
-   		}
-   		if(supplier_contract_type.indexOf("land")>-1){
-   			saveJobLandSpContractConditions(jobOrder);
-   		}
+   		
    		
    		//海 陆 空运带出合同费用信息(客户)
-   		if(customer_contract_type.indexOf("ocean")>-1){
-   			saveJobOceanCustomerContractConditions(jobOrder);
+   		if(customer_contract_type != null){
+   			if(customer_contract_type.indexOf("ocean")>-1){
+   	   			saveJobOceanCustomerContractConditions(jobOrder);
+   	   		}
+   	   		if(customer_contract_type.indexOf("air")>-1){
+   	   			saveJobAirCustomerContractConditions(jobOrder);
+   	   		}
+   	   		if(customer_contract_type.indexOf("land")>-1){
+   	   			saveJobLandCustomerContractConditions(jobOrder);
+   	   		}
    		}
-   		if(customer_contract_type.indexOf("air")>-1){
-   			saveJobAirCustomerContractConditions(jobOrder);
-   		}
-   		if(customer_contract_type.indexOf("land")>-1){
-   			saveJobLandCustomerContractConditions(jobOrder);
-   		}
+   		
 
    		//空运带出合同费用信息(供应商)
    		saveJobAirSpContractConditions(jobOrder);
@@ -3496,8 +3505,8 @@ public class JobOrderService extends Controller {
     	
     	Map map = new HashMap();
         map.put("sEcho", 1);
-        map.put("iTotalRecords", list.size());
-        map.put("iTotalDisplayRecords", list.size());
+        map.put("iTotalRecords", list != null?list.size():0);
+        map.put("iTotalDisplayRecords", list != null?list.size():0);
         map.put("aaData", list);
         renderJson(map); 
     }
@@ -3749,20 +3758,23 @@ public class JobOrderService extends Controller {
     	
         List<Map<String, String>> land_charge_item = (ArrayList<Map<String, String>>)dto.get("land_charge_item");
         Model<?> model = (Model<?>) JobOrderArap.class.newInstance();
-        for(int i=0;i<land_charge_item.size();i++){
-        	Map<String, String> map=land_charge_item.get(i);
-        	
-        	DbUtils.setModelValues(map,model);
-        	model.set("land_item_id", land_item_id);
-        	model.set("order_id", order_id);
-        	if("UPDATE".equals(map.get("action"))){
-        		model.update();
-        	}else if("DELETE".equals(map.get("action"))){
-        		model.delete();
-        	}else{
-        		model.save();
-        	}
+        if(land_charge_item != null){
+        	for(int i=0;i<land_charge_item.size();i++){
+            	Map<String, String> map=land_charge_item.get(i);
+            	
+            	DbUtils.setModelValues(map,model);
+            	model.set("land_item_id", land_item_id);
+            	model.set("order_id", order_id);
+            	if("UPDATE".equals(map.get("action"))){
+            		model.update();
+            	}else if("DELETE".equals(map.get("action"))){
+            		model.delete();
+            	}else{
+            		model.save();
+            	}
+            }
         }
+        
       //保存陆运费用模版
         String type = (String) dto.get("type");//根据工作单类型生成不同前缀
         String customer_id = (String)dto.get("customer_id");
